+++ src/components/Modal/index.tsx
@@ -11,26 +11,22 @@ type Props = {
 export function Modal({ children, title, className }: Props) {
   return (
     <Dialog.Portal>
-      <Dialog.Overlay
-        className={`
-          fixed left-0 top-0 flex h-screen w-screen items-center justify-center bg-zinc-900/30 
-          backdrop-blur-md ${className}
-        `}
+      <Dialog.Overlay className="fixed inset-0 bg-black/60 backdrop-blur-sm" />
+      <Dialog.Content
+        onCloseAutoFocus={(e) => e.preventDefault()}
+        className={`fixed left-1/2 top-1/2 -translate-x-1/2 -translate-y-1/2 flex flex-col max-h-[90vh] min-w-[200px] rounded-md bg-zinc-800 p-6 text-zinc-100 shadow-xl ${className ?? ""}`}
       >
-        <Dialog.Content className="min-w-[50vw] rounded-md bg-zinc-200 p-4 pb-2 text-zinc-900">
-          <header className="mb-8 flex justify-between">
-            <Dialog.Title className="font-bold">{title}</Dialog.Title>
+        <header className="mb-8 flex justify-between">
+          <Dialog.Title className="font-bold">{title}</Dialog.Title>
 
-            <Dialog.Close className="cursor-pointer">
-              <X className="text-red-400" size={18} weight="bold" />
-            </Dialog.Close>
-          </header>
+          <Dialog.Close className="cursor-pointer">
+            <X className="text-red-400" size={18} weight="bold" />
+          </Dialog.Close>
+        </header>
 
-          <Dialog.Description />
-          {children}
-          <Dialog.Close />
-        </Dialog.Content>
-      </Dialog.Overlay>
+        <Dialog.Description />
+        {children}
+      </Dialog.Content>
     </Dialog.Portal>
   );
 }