+++ src/contexts/TicketContext/actions.ts
@@ -4,6 +4,7 @@ export type FilterProps = {
   expiry_date?: Date;
   limite_expire_date?: Date;
   document_number?: string;
+  payment_place?: string;
   is_online?: boolean;
 };
 
@@ -11,5 +12,6 @@ export const defaultFilter: FilterProps = {
   recipient: "",
   type: "all",
   document_number: "",
+  payment_place: "",
   is_online: undefined,
 };